refactor: Move lexer implementation into mod.rs
Consolidates the lexer code by moving implementation from lr_lexer.rs
into the mod.rs file where it was originally located. This simplifies
the module structure and makes the code more canonical.
